pyproject.toml
[project]
name = "djangocon-cal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]

[project.scripts]
djangocon-cal = "main:main"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
]
